test: move the setup code of firebase-functions-test to index.helper.ts
